Move build_base_url logic into connections and add ssl.
Instead of building base url separately in handle and cursor, we now build base_url when opening the connection. This is more elegant and also enables us to specify if we want to use an encrypted protocol instead. Also changed profile_template.yml to reflect the fact we're using rest api instead of pyodbc.