Add script for manual tests (#37)
### Summary

Created a script to automate the manual smoke test we were doing before
merging PRs. Next step is to call this using a github workflow so that
we can automate testing as soon as a PR is created.

### Description

Created a script which takes profiles.yml options as parameters. Another
option is specified to decide if we're testing against software or
cloud. These parameters will be fixed in our workflow, so no need to
memorize the order of them to run the script manually.

### Related Issue

https://github.com/dremio/dbt-dremio/issues/35

### Additional Reviewers

@ArgusLi 
@jlarue26

Co-authored-by: ArgusLi <argus.li@dremio.com>