Create slightly too high up unit test.
Create unit test that stubs the _post method and returns DremioRequestTimeoutException(). The reason why this unit test is slightly too high up is because it does not test the _check_error() and _raise_for_status() logic. Instead, it tests the dbt_core base adapter retry_connection logic, which is not what it should test.