Get functional tests working against Dremio Cloud (#47)
### Summary

Made changes to the tests to refer to the S3 bucket instead of the GCS
bucket. The source name is still the same, but the root_path in the
tests will now prefix the bucket name, instead of the source name. In
addition, I've added a dremio cloud profile and dremio software profile
to run the functional tests against. Since the privileges table is
different in Dremio Software and Dremio Cloud, the grants macro needs to
check which profile we're using in order to get data from the privileges
table.

### Description

#### Grants Macro
- Check to see if we're using Dremio Cloud or Dremio Software as our
host and set the privileges_table name accordingly
- Quote the grantee user, as Dremio Cloud users have dots in their name

#### Functional Tests Configuration
- Add a new parser option for pytest called `--profile`
- Add two profile targets: dremio_cloud_target() and
dremio_software_target()

These changes allow us to specify 'dremio_cloud' or 'dremio_software' as
the profile when running the functional tests

#### Datalake Constant
- Rename to 'BUCKET' and add another constant called 'SOURCE' to
differentiate between the two in Dremio

### Related Issue

### Additional Reviewers

@jlarue26 
@ArgusLi