Create macro dremio__rename_relation that uses CTAS and DROP instead of default ALTER TABLE RENAME to.
Default rename_relation uses ALTER TABLE and RENAME to. Dremio does not currently support that. The new macro replaces that with 2 call statements - 1 to create a new table with the new name and old contents, 2 to drop the old table. 2 call statements are needed as Dremio only supports single statement execution.