Fix schema parsing (#205)
### Summary

Prior to this change, users would not be able to specify a schema with a
dot in it. For instance, schema.parse would get quoted as
"schema"."parse", even though the user intended to quote the entire
identifier. Now users can choose to wrap the schema in single quotes and
add double quotes around the parts they want.

### Description

Add regex to match a pattern containing quotes. Solution provided by
@dremio-brock.

### Test Results

All tests pass.

### Changelog

-   [x] Added a summary of what this PR accomplishes to CHANGELOG.md

### Related Issue

https://github.com/dremio/dbt-dremio/issues/203