Ensure adapter does not fetch actual records from materialization (#214)
### Summary

During the second run of an incremental materialization, the adapter
would try and fetch records from the temporary table it is creating.
This made the adapter hang if the table was large enough.

### Description
Made the fetch configuration false by default. Also changed the
run_query macro to use call_statement, as run_query will always fetch
the data that is returned by the sql query. Thanks to @mxmarg for the
initial iteration of this fix.

### Related Issue

https://github.com/dremio/dbt-dremio/issues/211