Make schema parsing backwards compatible (#216)
### Summary

We made changes to the schema parsing that allowed for schemas that
contained dots. However, this requires users to specify ALL schemas
wrapped in single quotes and then double quotes (`'"my_schema"'`),
otherwise it won't add any quotes to schemas. With this change users can
continue to add their schema as normal (`'my_schema'`).

### Description

Added a condition to check if double quotes are contained in the schema
definition. If so, we use the most recent logic. If not we use the
regular logic.

### Test Results

<!--- Details of the tests you've ran -->

### Changelog

-   [ ] Added a summary of what this PR accomplishes to CHANGELOG.md

### Related Issue

<!--- Link to issue where this is tracked -->