Move Reflections to Rest API Calls and Extend Implementation (#256)
### Summary

Reflections were being created through the SQL Runner and that was the
cause for some issues we were facing.

### Description

The current reflections implementation requires sys.reflections access,
which is only available to superusers (admins). When creating the
reflection with a regular user, it will error. To fix this issue, the
creation of the reflections has been moved from queries being ran inside
the SQL Runner to Dremio's Rest API calls.
The current logic of writing reflections through a custom
materialization remains to make sure reflections created under the
previous approach still work.

On top of this, the behaviour where all reflections would be dropped and
then recreated when running a dbt project has been changed to instead
first check if a reflection with the same name is already present in the
same dataset and, if that's the case, it will update the already
existing reflection instead.

### Test Results

A total of 11 tests have been added, approaching different situations
that may occur when dealing with reflections

### Changelog

- Reflections are now handled through the Rest API, allowing non-admin
users to also create them
- It is now possible to set a custom name for reflections
- If a reflection already exists in the dataset with the same name
defined in the model, it will be updated instead of creating a new one
- New `date_dimensions` parameter was added to the reflection
materialization, to set fields that have a `DATE` granularity
- Added distribution fields under `distribute_by`
- Added partition transformations under `partition_transform`
  - Defaults to Original/Identity if not defined
  - year/month/day/hour/bucket(n), truncate(n)
- Computations default to `SUM, COUNT` if mapped measure is numeric,
`COUNT` if not
- `reflections_enabled` adapter option has been renamed to
`reflections_metadata_enabled` (requires user privileges to run in
dremio)
- CI now creates and formats a `Samples` source as it is needed for the
reflection tests