added walkthough (#265)
### Summary

Wrote a Walkthrough that covers most of the dbt basic and adapter
features and nuances with links to other resources on features not
covered. Also added a link to this walkthrough from main readme.md

### Description

Wrote a Walkthrough that covers most of the dbt basic and adapter
features and nuances with links to other resources on features not
covered. Also added a link to this walkthrough from main readme.md

### Test Results

Wrote a Walkthrough that covers most of the dbt basic and adapter
features and nuances with links to other resources on features not
covered. Also added a link to this walkthrough from main readme.md

### Changelog

Just Docs

### Related Issue

No Issue