Support for roles in grants (#259)
### Summary

Added support for roles in grants

### Description

Grants can now be set for both users and for roles.
A prefix was added to handle this, with `user:` and `role:` being the
valid prefixes. For example, `user:dbt_test_user_1` and
`role:dbt_test_role_1`.
If no prefix is provided, defaults to user for backwards compatibility.

### Test Results

Tests were ran against a local Dremio Enterprise instance and are
skipped since the current pipeline uses Dremio OSS, and grants are an
Enterprise feature

### Changelog

- Added support for roles in grants

### Related Issue

#177