Fix twin strategy implementation (#278)
### Summary

dbt assumes tables and views share locations, unlike Dremio that keeps
tables and views separately. We address this conflict using the
[twin-strategy](https://github.com/dremio/dbt-dremio/wiki/Using-Materializations-with-Dremio#user-content-optional-twin-strategy-configuration).
Currently, the adapter's clone strategy seems to create a view for every
table, even if there is no conflicting view to be replaced.

### Description

With this change, we introduce a check that will only overwrite existing
views, not create new ones on table creation when the `clone` strategy
is enabled. When a view is created with a conflicting table, the view
will always select the existing table.

### Test Results

Due to this change, tests that queried the corresponding view after
table creation (instead of the table directly) were failing due to this
change. The schema of all affected tests have been adjusted.

### Changelog

- [x] Only overwrite existing views in case of schema and name conflict
- [x] Adjust tests to query created tables directly instead of their
matching views
-   [x] Added a summary of what this PR accomplishes to CHANGELOG.md