Ability to perform unit tests on models with dependencies - fixing ref() macro (#293)
### Summary

Fixing our custom `ref()` macro in order to support unit testing on
models with dependencies.

### Description

This is originated from the error we were getting while trying to
perform unit tests on models with dependencies → `Compilation Error ...
'dict object' has no attribute 'nodes'`. Our original override of
`ref()` macro relies on `graph` Jinja function in some specific cases to
retrieve the metadata of a model.
However, this `graph` is not available (with content) during unit tests,
but it's essential for other cases (e.g. **reflection** models).
Therefore, we added a check before its usage and enabled `ref()` on unit
testing models. With this fix, we basically are using the builtin
`ref()`'s result.

### Test Results

- Refactored tests for unit testing
- All tests are passing

### Changelog

-   [x] Added a summary of what this PR accomplishes to CHANGELOG.md