Fix data type issue in fixtures creation (#294)
### Summary

Fixes data type casting issues while mocking unit testing data.

### Description

We noticed that not all data types were being correctly parsed from dbt
to Dremio in fixtures creation. This was happening because:
- dbt assumes all castings have the format `cast <data> as <type>` but
for Dremio SQL dialect this won't always work (e.g. arrays) → therefore,
we overrode `safe_cast` macro and adjusted the format of each casting
according to the type / context of the query.
- The `input` field we provide in unit tests configuration will
internally become an ephemeral model built based on the model being
tested. This input model creation relies on the metadata stored in
Dremio - column names, types, etc -, so in order to do this mapping from
Dremio results to a table we can manipulate in dbt's environment,
`agate` library was used. The problem is this library **infers** the
types of the column values in a specific order by default
([agate.TypeTester](https://agate.readthedocs.io/en/latest/api/type_tester.html))
which led to the numeric type casting issues (values like _0_, _1_ and
_null_ were all being treated as **Boolean**) → the fix is to force it
to use a specific data type in types where the values that can be
misinterpreted as Boolean.

### Test Results

- Refactored unit testing data types test → based on Dremio supported
types, added tests for them
- All tests are passing

### Changelog

-   [x] Added a summary of what this PR accomplishes to CHANGELOG.md